requirements.txt
openai
langchain>=0.1.17
langchain-community>=0.0.19
chromadb
python-dotenv
PyPDF2
python-docx
tqdm
tiktoken
langchain-openai>=0.0.8
langchain-chroma>=0.0.8
langchain-core>=0.1.47
pypdf




